+++ src/zipline/data/resample.py
@@ -79,12 +79,12 @@ def minute_to_session(column, close_locs, data, out):
         The `open`, `high`, `low`, `close`, or `volume` column.
     close_locs : array[intp]
         The locations in `data` which are the market close minutes.
-    data : array[float64|int64]
+    data : array[float64|uint32]
         The minute data to be sampled into session data.
         The first value should align with the market open of the first session,
         containing values for all minutes for all sessions. With the last value
         being the market close of the last session.
-    out : array[float64|int64]
+    out : array[float64|uint32]
         The output array into which to write the sampled sessions.
     """
     if column == "open":
@@ -538,7 +538,7 @@ class MinuteResampleSessionBarReader(SessionBarReader):
             if col != "volume":
                 out = np.full(shape, np.nan)
             else:
-                out = np.zeros(shape, dtype=np.int64)
+                out = np.zeros(shape, dtype=np.uint32)
             results.append(out)
 
         for i in range(len(assets)):
@@ -688,7 +688,7 @@ class ReindexBarReader(ABC):
             if field != "volume":
                 out = np.full(shape, np.nan)
             else:
-                out = np.zeros(shape, dtype=np.int64)
+                out = np.zeros(shape, dtype=np.uint32)
 
             if inner_results is not None:
                 out[indices] = inner_results[i]
