+++ tests/utils/daily_bar_writer.py
@@ -5,7 +5,7 @@ from numpy import (
 )
 from bcolz import ctable
 
-from zipline.data.bcolz_daily_bars import BcolzDailyBarWriter, OHLC, INT64_MAX
+from zipline.data.bcolz_daily_bars import BcolzDailyBarWriter, OHLC, UINT32_MAX
 
 
 class DailyBarWriterFromDataFrames(BcolzDailyBarWriter):
@@ -28,18 +28,18 @@ class DailyBarWriterFromDataFrames(BcolzDailyBarWriter):
         arrmax = array.max()
         if colname in OHLC:
             self.check_uint_safe(arrmax * 1000, colname)
-            return (array * 1000).astype(int64)
+            return (array * 1000).astype(uint32)
         elif colname == "volume":
             self.check_uint_safe(arrmax, colname)
-            return array.astype(int64)
+            return array.astype(uint32)
         elif colname == "day":
             nanos_per_second = 1000 * 1000 * 1000
             self.check_uint_safe(arrmax.view(int64) / nanos_per_second, colname)
-            return (array.view(int64) / nanos_per_second).astype(int64)
+            return (array.view(int64) / nanos_per_second).astype(uint32)
 
     @staticmethod
     def check_uint_safe(value, colname):
-        if value >= INT64_MAX:
+        if value >= UINT32_MAX:
             raise ValueError(
                 "Value %s from column '%s' is too large" % (value, colname)
             )
